refactor: renamed js files to jsx
