+++ src/renderer/components/NavBar.jsx
