+++ src/renderer/components/PdfDisplay.jsx
