+++ src/renderer/components/microComponents/Selector.jsx
