+++ src/renderer/pages/ApplicationPage/ApplicationDetails.jsx
