+++ src/renderer/pages/ApplicationPage/CVSection/CVConstructorSection/CVConstructorSection.jsx
