+++ src/renderer/pages/ApplicationPage/CVSection/CVConstructorSection/IndividualFormElement.jsx
