+++ src/renderer/pages/ApplicationPage/CVSection/CVConstructorSection/NewCVSectionForm.jsx
