+++ src/renderer/pages/ApplicationPage/CVSection/CVSection.jsx
