+++ src/renderer/pages/ApplicationPage/LetterSection/LetterConstructorSection.jsx
