+++ src/renderer/pages/ApplicationSummaryPage/DeleteComponentModal.jsx
